src/main/java/com/udacity/jwdnd/course1/cloudstorage/services/UserService.java
package com.udacity.jwdnd.course1.cloudstorage.services;

import com.udacity.jwdnd.course1.cloudstorage.mapper.UserMapper;
import com.udacity.jwdnd.course1.cloudstorage.model.User;
import org.springframework.dao.DuplicateKeyException;
import org.springframework.stereotype.Service;

import java.security.SecureRandom;
import java.util.Base64;

@Service
public class UserService {

    private final UserMapper userMapper;
    private final HashService hashService;

    public UserService(UserMapper userMapper, HashService hashService) {
        this.userMapper = userMapper;
        this.hashService = hashService;
    }

    public boolean isUsernameAvailable(String username) {
        return userMapper.getUser(username) == null;
    }

    public int createUser(User user) {
        SecureRandom random = new SecureRandom();
        byte[] salt = new byte[16];
        random.nextBytes(salt);
        String encodedSalt = Base64.getEncoder().encodeToString(salt);
        String hashedPassword = hashService.getHashedValue(user.getPassword(), encodedSalt);
        int result = 0;
        try {
            result = userMapper.insert(new User(null, user.getUsername(), encodedSalt, hashedPassword, user.getFirstName(), user.getLastName()));
        } catch (DuplicateKeyException e) {
            // @todo: replace System.out using a logging service to log the error message.
            System.out.println("Username '" + user.getUsername() + "' already exists.");
        }
        return result;
    }

    public User getUser(String username) {
        return userMapper.getUser(username);
    }

    public int deleteUser(User user) {
        return userMapper.delete(user);
    }

    public int deleteAllUsers() {
        return userMapper.deleteAll();
    }
}
